@@ -24,5 +24,5 @@ class Admin < ApplicationRecord
   # :confirmable, :lockable, :timeoutable and :omniauthable
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :trackable, :validatable
-  has_many :articles, as: :commentable , :dependent => :destroy       
+  has_many :articles, as: :commentable , :dependent => :destroy   
 end
