@@ -11,5 +11,6 @@ class PageChannel < ApplicationCable::Channel
     ActionCable.server.broadcast 'page_channel', message: data['message']
     Comment.create! commenter:current_admin.user_name , body: data['message'], commentable_type: "Article",  commentable_id:1
   end
-  
+
 end
+
