@@ -12,7 +12,8 @@
 #
 
 class Comment < ApplicationRecord
-    belongs_to :commentable, polymorphic: true 
+    belongs_to :commentable, polymorphic: true, optional: true 
     #validates_associated :commentable
     validates :body, presence: true, length: {minimum: 5, maximimum: 1000 }
+    after_create_commit { CommentBroadcastJob.perform_later(self) }
 end
