@@ -1,5 +1,11 @@
-Warden::Manager.after_set_user do |admin,auth,opts|
+Warden::Manager.after_set_user do |user,auth,opts|
   scope = opts[:scope]
-  auth.cookies.signed["#{scope}.id"] = admin.id
-  auth.cookies.signed["#{scope}.expires_at"] = 30.minutes.from_now
+  auth.cookies.signed["#{scope}.id"] = user.id
+  auth.cookies.signed["#{scope}.expires_at"] = 60.minutes.from_now
+end
+
+Warden::Manager.before_logout do |user, auth, opts|
+  scope = opts[:scope]
+  auth.cookies.signed["#{scope}.id"] = nil
+  auth.cookies.signed["#{scope}.expires_at"] = nil
 end
